更新登录、登出、查询登录用户信息

+++ pluskynet/src/com/pluskynet/dao/impl/SampleDaoImpl.java
@@ -13,6 +13,7 @@ import com.pluskynet.domain.Article01;
 import com.pluskynet.domain.Articleyl;
 import com.pluskynet.domain.Rule;
 import com.pluskynet.domain.Sample;
+import com.pluskynet.domain.User;
 import com.pluskynet.service.SampleService;
 
 public class SampleDaoImpl extends HibernateDaoSupport implements SampleDao{
@@ -22,35 +23,37 @@ public class SampleDaoImpl extends HibernateDaoSupport implements SampleDao{
 	@Transactional
 	public List<Article01> getListArticle(String table, String year, int count,String trialRound,String doctype) {
 		String hql = "SELECT * FROM "+table+" WHERE `id`  >= ((SELECT MAX(id) FROM "+table+")-(SELECT MIN(id) FROM "+table+")) * RAND() + (SELECT MIN(id) FROM "+table+") "
-				+ "and `date` ='"+year+"' and title like '%"+trialRound+"%' and title like '%"+doctype+"%' limit 1000";
+				+ "and `date` ='"+year+"' and spcx ='"+trialRound+"' and doctype = '"+doctype+"' limit "+count+"";
 		Session session = this.getHibernateTemplate().getSessionFactory().getCurrentSession();
 		List<Article01> list = session.createSQLQuery(hql).addEntity(Article01.class).list();
 		return list;
 	}
 
 	@Override
-	public void save(List<Article01> list) {
+	public void save(List<Article01> list,User user) {
 		for (int i = 0; i < list.size(); i++) {
 			Articleyl articleyl = new Articleyl();
 			articleyl.setDecodeData(list.get(i).getDecodeData());
 			articleyl.setDocId(list.get(i).getDocId());
 			articleyl.setTitle(list.get(i).getTitle());
+			articleyl.setBelongid(user.getUserid());
+			articleyl.setBelonguser(user.getUsername());
 			this.getHibernateTemplate().save(articleyl);
 		}
 	}
 
 	@Override
 	@Transactional
-	public void saverule(Sample sample) {
-		String hql = "delete From Sample";
+	public void saverule(Sample sample,User user) {
+		String hql = "delete From Sample where userid = "+user.getUserid()+"";
 		Session session = this.getHibernateTemplate().getSessionFactory().getCurrentSession();
 		session.createSQLQuery(hql).executeUpdate();
 		this.getHibernateTemplate().save(sample);
 	}
 
 	@Override
-	public Sample select() {
-		String hql = "from Sample";
+	public Sample select(User user) {
+		String hql = "from Sample where userid = "+user.getUserid()+"";
 		List<Sample> list = this.getHibernateTemplate().find(hql);
 		if (list.size()>0) {
 			return list.get(0);
@@ -61,8 +64,8 @@ public class SampleDaoImpl extends HibernateDaoSupport implements SampleDao{
 
 	@Override
 	@Transactional
-	public void delete() {
-		String hql = "delete From Articleyl";
+	public void delete(User user) {
+		String hql = "delete From Articleyl where userid = "+user.getUserid()+"";
 		Session session = this.getHibernateTemplate().getSessionFactory().getCurrentSession();
 		session.createSQLQuery(hql).executeUpdate();
